+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * Add `#[WhenNot]` attribute to prevent service from being registered in a specific environment
  * Generate a meta file in JSON format for resource tracking
  * Add `SkippingResourceChecker`
+ * Add support for `defaultNull()` on `BooleanNode`
 
 7.1
 ---